Merge GitHub reporting repository defaults with custom GitHub reporting configuration
Report GitHub configuration is merged with the default GitHub reporting configuration from the action context:
```
github:
  username: GITHUB_ACTOR,
  owner: GITHUB_REPOSITORY_OWNER,
  token,
  "project-name": GITHUB_REPOSITORY,
```

Before the default github reporting function was used and did not
take the custom configuration into account. Especially the duplication
detection and the severity as label was not used when set the github-report flag.

Signed-off-by: fisehara <harald@balena.io>